Split out UI to separate file.
